requirements.txt
Django==4.2.7
mysql-connector-python==8.3.0
python-decouple==3.8
requests==2.31.0
openai==1.3.5
gunicorn==21.2.0
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
